+++ Godunov/hydro_godunov.cpp
@@ -38,7 +38,7 @@ Godunov::ComputeAofs ( MultiFab& aofs, const int aofs_comp, const int ncomp,
     // if we need convetive form, we must also compute
     // div(u_mac)
     MultiFab divu_mac(state.boxArray(),state.DistributionMap(),1,4);;
-    for (int i = 0; i < iconserv.size(); ++i)
+    for (long unsigned i = 0; i < iconserv.size(); ++i)
     {
         if (!iconserv[i])
         {
